Move SubmitImpl and initialize methods to base clase GXBaseObject.
