+++ dotnet/src/dotnetframework/GxClasses/Middleware/GXHttp.cs
@@ -259,7 +259,7 @@ namespace GeneXus.Http
 		public virtual void InitializeDynEvents() { throw new Exception("The method or operation is not implemented."); }
 		public virtual void initialize_properties() { throw new Exception("The method or operation is not implemented."); }
 		public virtual void webExecute() { throw new Exception("The method or operation is not implemented."); }
-		public virtual void initialize() { throw new Exception("The method or operation is not implemented."); }
+		
 		public virtual bool SupportAjaxEvent() { return false; }
 		public virtual String AjaxOnSessionTimeout() { return "Ignore"; }
 #if NETCORE
