+++ dotnet/src/dotnetframework/GxClasses/Model/GXWebProcedure.cs
@@ -10,6 +10,7 @@ namespace GeneXus.Procedure
 	using GeneXus.Utils;
 	using log4net;
 	using GeneXus.Application;
+	using GeneXus.Data.NTier;
 
 	public class GXWebProcedure : GXHttpHandler
 	{
@@ -37,12 +38,6 @@ namespace GeneXus.Procedure
 		public override void initialize() { }
 		protected override void createObjects() { }
 		public override void skipLines(long nToSkip) { }
-		protected void SubmitImpl()
-		{
-			context.SetSubmitInitialConfig(context);
-			initialize();
-			Submit(ExecutePrivateCatch, this);
-		}
 		public override void cleanup()
 		{
 			if (!context.WillRedirect() && context.IsLocalStorageSupported())
